+++ changelog.md
@@ -3,6 +3,8 @@ All notable changes after version 1.6.16 to this project will be documented in t
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ##[unreleased] Unreleased
+### Added
+- the `$sleep` parameter to the `WPDb::haveMultisiteInDatabase` method
 
 ##[1.9.1] 2016-01-7
 ### Fixed